feat: add time on judge page
